Validations Buy Product

+++ src/app/components/product-list/product-list.component.ts
@@ -2,9 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { ProductService } from 'src/app/services/product.service';
 import { Product } from 'src/app/models/product.model';
 import { Router } from '@angular/router';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { AuthService } from 'src/app/services/auth.service';
 import { RouteConstants } from 'src/app/utils/route-constants';
+
 
 @Component({
   selector: 'app-product-list',
@@ -15,7 +15,7 @@ export class ProductListComponent implements OnInit {
 
   listProducts: Product[];
   typeUser: any;
-  constructor(private productService: ProductService, private router: Router, private _modalService: NgbModal,
+  constructor(private productService: ProductService, private router: Router,
     private authService: AuthService) {
     this.listProducts = [];
     this.getProductList();
@@ -25,14 +25,16 @@ export class ProductListComponent implements OnInit {
 
   }
 
-
+  /**
+   * 
+   */
   getProductList() {
 
     this.productService.getProductService().subscribe((res: any) => {
 
       this.listProducts = [];
       res.forEach((element: Product) => {
-        this.listProducts.push(new Product(element.name, element.description, element.value, element.ammount, element.id))
+        this.listProducts.push(new Product(element.name, element.description, element.value, 0, element.id))
       });
       setTimeout(() => {
         this.typeUser = this.authService.typeUser;
@@ -40,16 +42,27 @@ export class ProductListComponent implements OnInit {
     })
   }
 
+  /**
+   * 
+   * @param item 
+   */
   buyProduct(item) {
 
+
     if (JSON.parse(localStorage.getItem('autenticated'))) {
-      console.log(JSON.stringify(item))
+      const r = confirm('Esta seguro de comprar este producto');
+      if (r) {
+        console.log(JSON.stringify(item))
+      }
     } else {
       alert('Debes Ingresar para realizar la compra');
       this.router.navigate(['login']);
     }
   }
 
+  /**
+   * 
+   */
   navigateToCreateProduct() {
     this.router.navigate([RouteConstants.CREATE_PRODUCT_PATH]);
   }
@@ -67,10 +80,27 @@ export class ProductListComponent implements OnInit {
    * 
    */
   deleteProduct(item: Product) {
-    this.productService.deleteProduct(item.id).subscribe(res => {
-      alert('se ha eliminado con exito');
-      this.getProductList();
-    });
+    const r = confirm('Desea eliminar el producto?');
+    if (r) {
+      this.productService.deleteProduct(item.id).subscribe(res => {
+        alert('se ha eliminado con exito');
+        this.getProductList();
+      });
+    }
+  }
+
+  /**
+   * 
+   * @param val 
+   * @param item 
+   */
+  changeAmmount(val, item: Product) {
+
+    if (val === 'add') {
+      item.ammount = item.ammount + 1;
+    } else {
+      item.ammount = item.ammount - 1;
+    }
   }
 
 }
